Added Buttons
Buttons now work as expected, the button remains unhighlighted until the
cursor is hovering over it. The toolbar is now it's own button rather
than an invisible region of the window, this means that all the buttons
we lay on top of the toolbar will prevent us from being able to drag the
window which is a good thing. Added additional helper functions for
creating basic buttons. Reworked a lot of the unnessary code from the
StateMachine into a UIManager class.