Text Fields
Began implementing text fields. Added my own custom font that works
fairly well but still needs some adjusting. Need to make the text field
on button click rather than it being there 100% of the time. Then when
it loses focus we want to hide it. (As far as I am aware, it becomes
'unfocused' when you click off the text field. So I should be able to
just check if it is in focus or not to hide it, then unhide when we
click the button).