Lists
Changed the way that I was editing the Renderers for the widgets. Now
instead of editing the properties of each individual widget instance and
copy/pasting a bunch of code, I have just created a widget that is
edited to look the way I want it, then whenever creating more widgets of
the same type I just copy() them. Also got the list working so that it
display text correctly, now I need to begin working on having the text
stored in XML and read it into the list.