XML Loading
It now correctly loads XML data, for now it is just a basic int and
string that represents the characters name and ID. This data is then
used to fill the character list so it now display all the characters
found within the XML file. Clicking on a character in the list will
correctly display it's name in a text field that can be edited. Next I
need to work on getting it so that editing the name in this text field
will actually change the data within the XML file so on subsequent loads
it will show the new name.