Issues with Buttons hiding/showing
I am having some issues with the Search bar, the Search bar and the Name
text field both use the same functions and should function in the same
way but when I leave to another tab and return, the Search bar is not
being reset in the same way as the Name text field. This reset is
causing the Search bar to break when switching back to the Player tab. I
am creating a new branch where I will be removing the setOpacity() and
hide() functions to figure out what is going wrong.