Hopefully it is fixed now.
Seems to be working now, got rid of a lot of useless setOpacity() calls.
It now just keeps track if some text is entered into the search bar, if
there is then when we call Resume() we don't show() the image anymore. I
don't expect this to be the last of the issues though, will probably
creep up in the future.