Class Loading
Began working on loading class data when a class is selected from the
list. This is unfinished and not working properly yet, I have decided to
commit these changes to Git because I want to test something that
requires a small rework of how I am showing and hiding all the different
UI elements. Right now it is getting pretty messy having to hide() and
show() each individual element and it will only get worse, so I am going
to look into creating a parent UI element that will have it's own
children, hopefully this will allow me to make one single hide() or
show() call in order to hide all the UI elements I want.