Class Loading cont.
Unfortunately the previous plan didn't go exactly as planned. The reason
I decided to try and rework the way I was hiding the UI elements was
because I found some new classes in TGUI called containers that I though
were literally just containers for UI widgets, and I would be able to
call hide() once on the container to hide all of it's children, but it
doesn't actually work like that. However I have reworked it a little
bit, all of the setOpacity() calls are gone now as they were just over
complicating things, and I added some HideUI() functions that just make
it look a little cleaner when I need to hide an entire chunk of the UI.
Aside from that I am now loading all the character's classes from the
XML file correctly, and this information is being added to the text
fields correctly, next I need to work on reading the statistics of the
class and displaying those values.