Saving
Added saving functionality to the class name. Now when you change a
classes name it will correctly write the new name to the XML file, it
also edits the stored data within the DataManager so that we can update
the data whilst the program is still running. Reloading the class will
show the newly saved name and restarting the program will load the newly
saved name correctly.